+++ src/pyobsplot/static/widget.js
@@ -68338,9 +68338,7 @@ function generate_plot(spec, renderer) {
       out = out.plot();
     } else {
       if (spec["debug"]) {
-        console.log("--- start pyobsplot debugging output ---");
-        console.log(out);
-        console.log("--- end pyobsplot debugging output ---");
+        debug_output(out, renderer);
       }
       out = plot2(out);
     }
@@ -68351,6 +68349,18 @@ function generate_plot(spec, renderer) {
     out.textContent = "\u26A0 " + error;
   }
   return out;
+}
+function debug_output(out, renderer) {
+  if (renderer == "widget") {
+    console.log("--- start pyobsplot debugging output ---");
+    console.log(out);
+    console.log("--- end pyobsplot debugging output ---");
+  }
+  if (renderer == "jsdom") {
+    console.log("<br>--- start pyobsplot debugging output ---<br>");
+    console.log(out);
+    console.log("<br>--- end pyobsplot debugging output ---</br>");
+  }
 }
 
 // js/widget/widget.js