+++ NEWS.md
@@ -1,7 +1,7 @@
 ## pyobsplot (development version)
 
 - Autocompletion of Plot methods should now be working in IDEs.
-
+- debug mode also works with `jsdom` renderer.
 
 ## pyobsplot 0.3.2